Process multiple resource elements and check for a targetPath element
The targetPath element renames the directory in the loose application xml.
Since the processing of the configuration is so specific to the Maven War plugin, make a method dedicated to this.
Update and streamline the test case to check for the extra entry and remove the unnecessary test directory and Maven resource plugin.